Refactor: example raft-kv-rocksdb reuses example/rocksstore
