+++ examples/rocksstore/Cargo.toml
@@ -21,7 +21,7 @@ repository = "https://github.com/databendlabs/openraft"
 openraft = { path= "../../openraft", version = "0.10.0", features=["serde", "type-alias"] }
 
 rocksdb = "0.22.0"
-rand = "*"
+rand = "0.8"
 byteorder = "1.4.3"
 
 serde = { version = "1.0.114", features = ["derive"] }
